requirements.txt
# AutoPwnGPT Requirements

# Author: Eshan Roy
# Email: m.eshanized@gmail.com
# GitHub: https://github.com/TonmoyInfrastructureVision
# Date: 2025-04-22

# Core dependencies
PyQt6>=6.6.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
apscheduler>=3.10.1
# python-term>=1.0.0
jinja2>=3.1.2
markdown>=3.5
tqdm>=4.66.1
colorama>=0.4.6
rich>=13.6.0
pydantic>=2.5.0

# Network and security tools
aiohttp>=3.8.6
asyncio>=3.4.3
xmltodict>=0.13.0
beautifulsoup4>=4.12.2
urllib3>=2.0.7

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# Documentation
sphinx>=7.1.1
sphinx-rtd-theme>=1.3.0

# Optional dependencies (uncomment as needed)
# sqlalchemy>=2.0.0  # For database support
# fastapi>=0.104.1  # For API support
# uvicorn>=0.24.0   # For API support
# cryptography>=41.0.0  # For cryptographic operations
